Fix spelling error, add fucnction body to ret functions in database for equipment

+++ Source/Game/Zone.cpp
@@ -5,11 +5,11 @@
 
 Zone::Zone()
 {
-    m_background.setSize({Display::get().getSize().x,
-                          Display::get().getSize().y});
+    m_background.setSize({(float)Display::get().getSize().x,
+                          (float)Display::get().getSize().y});
 
-    m_shakeGround.setSize({Display::get().getSize().x,
-                           Display::get().getSize().y});
+    m_shakeGround.setSize({(float)Display::get().getSize().x,
+                           (float)Display::get().getSize().y});
 
     m_background  .setTexture(&getResources().getTexture(Texture_ID::Zone_Village));
     m_shakeGround .setTexture(&getResources().getTexture(Texture_ID::Zone_Village));
